+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feature `use-async-trait` renamed to `use-boxed-futures`
 - Feature `use-associated-future` renamed to `use-associated-futures`
 - Upgrade unimock to 0.4
+- New entrait option `mock_api` to optionally generate a mock setup interface.
 - Unimock mocks are now exported through a `TraitMock` ident, where `Trait` is the name of the trait.
 
 ## [0.4.6] - 2022-08-17